Added Chip types
